refactor(core): group option and runtime helpers
Move page range parsing and parallel execution helpers into focused option and runtime folders so the core root stays limited to the processor entrypoint.
