+++ src/core/options/pageRange.ts
